pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "kuberoscli"
version = "0.0.1"
authors = [
  { name="Yongzhou Zhang", email="yongzhou.zhang@h-ka.de" },
]
description = "KuebeROS Command Line Tool"
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.8"

classifiers = [
    "Programming Language :: Python :: 3",
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

keywords = ['robotics', 'ros', 'kuberos']

dependencies = [
    'tabulate >= 0.9.0',
    'requests >= 2.29.0',
    'argcomplete >= 3.0.0',
    'ruamel.yaml >= 0.17.4',
    'ruamel.yaml.clib >= 0.2.6',
    'PyYAML >= 6.0.0',
]

[project.scripts]
kuberos = "kuberoscli.kuberoscli:main"

[tool.setuptools]
include-package-data = true
packages = ['kuberoscli', 'kuberoscli.command_group']

[project.urls]
"Homepage" = "https://github.com/kuberos-io/kuberos-cli"
"Bug Tracker" = "https://github.com/kuberos-io/kuberos-cli/issues"
